Separate out Builder and add Tabs with JSON importer
