+++ app/Filament/Forms/Resources/FormVersionResource.php
@@ -2,26 +2,23 @@
 
 namespace App\Filament\Forms\Resources;
 
-use App\Filament\Components\ContainerBlock;
-use App\Filament\Components\FieldGroupBlock;
-use App\Filament\Components\FormFieldBlock;
+use App\Filament\Components\FormVersionBuilder;
 use App\Filament\Forms\Resources\FormVersionResource\Pages;
-use App\Helpers\FormTemplateHelper;
+use App\Helpers\ImportTemplateHelper;
+use App\Helpers\ScanTemplateHelper;
 use App\Models\FormVersion;
 use Filament\Forms\Form;
 use Filament\Resources\Resource;
 use Filament\Tables;
 use Filament\Tables\Table;
 use Illuminate\Support\Facades\Gate;
-use Filament\Forms\Components\Select;
-use Filament\Forms\Components\TextInput;
 use Filament\Forms\Components\Textarea;
-use Filament\Forms\Components\Fieldset;
-use Filament\Forms\Components\DateTimePicker;
-use Filament\Forms\Components\Section;
 use Filament\Forms\Components\Actions;
 use Filament\Forms\Components\Actions\Action;
-use Filament\Forms\Components\Builder;
+use Filament\Forms\Components\MarkdownEditor;
+use Filament\Forms\Components\Split;
+use Filament\Forms\Components\Tabs;
+use Filament\Forms\Components\Tabs\Tab;
 use Filament\Forms\Get;
 use Filament\Forms\Set;
 use Illuminate\Support\Facades\Session;
@@ -47,140 +44,65 @@ class FormVersionResource extends Resource
         Session::put('elementCounter', $currentCounter + 1);
     }
 
-    // Function to find highest used instance ID
-    protected static function getHighestID(array $blocks): int
-    {
-        $maxID = 0;
-        foreach ($blocks as $block) {
-            // Check top-level elements
-            if (isset($block['data']['instance_id'])) {
-                $idString = $block['data']['instance_id'];
-                $numericPart = str_replace('element', '', $idString); // Remove the 'element' prefix
-                if (is_numeric($numericPart) && $numericPart > 0) {
-                    $id = (int) $numericPart;
-                    $maxID = max($maxID, $id); // Update the maximum ID
-                }
-            }
-            // Recursively check elements inside of Containers
-            if (isset($block['data']['components']) && is_array($block['data']['components'])) {
-                $nestedMaxID = self::getHighestID($block['data']['components']);
-                $maxID = max($maxID, $nestedMaxID);
-            }
-            // Recursively check elements inside of Groups
-            if (isset($block['data']['form_fields']) && is_array($block['data']['form_fields'])) {
-                $nestedMaxID = self::getHighestID($block['data']['form_fields']);
-                $maxID = max($maxID, $nestedMaxID);
-            }
-        }
-
-        return $maxID;
-    }
-
     public static function form(Form $form): Form
     {
         return $form
             ->schema([
-                Select::make('form_id')
-                    ->relationship('form', 'form_id_title')
-                    ->required()
+                // Main view
+                FormVersionBuilder::schema()
+                    ->visible(fn($livewire) => !($livewire instanceof \Filament\Resources\Pages\CreateRecord)),
+
+                // Tab view for Create page
+                Tabs::make('Tabs')
+                    ->visible(fn($livewire) => ($livewire instanceof \Filament\Resources\Pages\CreateRecord))
+                    ->columnSpanFull()
+                    ->activeTab(2)
                     ->reactive()
-                    ->preload()
-                    ->searchable()
-                    ->default(request()->query('form_id_title')),
-                Select::make('status')
-                    ->options([
-                        'draft' => 'Draft',
-                        'testing' => 'Testing',
-                        'archived' => 'Archived',
-                        'published' => 'Published',
-                    ])
-                    ->required(),
-                Section::make('Form Properties')
-                    ->collapsible()
-                    ->collapsed()
-                    ->columns(3)
-                    ->compact()
-                    ->schema([
-                        Fieldset::make('Requester Information')
+                    ->tabs([
+                        Tab::make('Build')
                             ->schema([
-                                TextInput::make('form_requester_name')
-                                    ->label('Name'),
-                                TextInput::make('form_requester_email')
-                                    ->label('Email')
-                                    ->email(),
-                            ])
-                            ->label('Requester Information'),
-                        Fieldset::make('Approver Information')
-                            ->schema([
-                                TextInput::make('form_approver_name')
-                                    ->label('Name'),
-                                TextInput::make('form_approver_email')
-                                    ->label('Email')
-                                    ->email(),
-                            ])
-                            ->label('Approver Information'),
-                        Select::make('deployed_to')
-                            ->label('Deployed To')
-                            ->options([
-                                'dev' => 'Development',
-                                'test' => 'Testing',
-                                'prod' => 'Production',
-                            ])
-                            ->columnSpan(1)
-                            ->nullable()
-                            ->afterStateUpdated(fn(callable $set) => $set('deployed_at', now())),
-                        DateTimePicker::make('deployed_at')
-                            ->label('Deployment Date')
-                            ->columnSpan(1),
-                        Select::make('form_data_sources')
-                            ->multiple()
-                            ->preload()
-                            ->columnSpan(1)
-                            ->relationship('formDataSources', 'name'),
-                        Textarea::make('comments')
-                            ->label('Comments')
+                                FormVersionBuilder::schema(),
+                            ]),
+                        Tab::make('Import')
                             ->columnSpanFull()
-                            ->maxLength(500),
+                            ->schema([
+                                Split::make([
+                                    Textarea::make('json')
+                                        ->label('Import JSON')
+                                        ->afterStateUpdated(fn(Set $set) => $set('jsonModified', true))
+                                        ->rows(15),
+                                    MarkdownEditor::make('messages')
+                                        ->label('Validation messages')
+                                        ->disabled(),
+                                ]),
+                                Actions::make([
+                                    Action::make('Scan JSON')
+                                        ->label('Scan JSON')
+                                        ->action(function (Set $set, $state) {
+                                            $validation = ScanTemplateHelper::validateForm($state['json']);
+                                            $set('messages', $validation['messages']);
+                                            $set('isValid', $validation['isValid']);
+                                            $set('jsonModified', false);
+                                        }),
+                                    Action::make('Import JSON')
+                                        ->label('Import JSON')
+                                        ->disabled(fn(Get $get) => !$get('isValid') || $get('jsonModified'))
+                                        ->action(function (Set $set, $state, $livewire) {
+                                            $formVersion = ImportTemplateHelper::importForm($state['json']);
+                                            if ($formVersion->id) {
+                                                $set('formVersion', $formVersion);
+                                            }
+                                        }),
+                                    Action::make('View Imported Form')
+                                        ->label('View Imported Form')
+                                        ->disabled(fn(Get $get) => !$get('formVersion')) // Only enable when formVersionId is set
+                                        ->action(
+                                            fn(Get $get, $livewire) =>
+                                            $livewire->redirect(FormVersionResource::getUrl('view', ['record' => $get('formVersion')]))
+                                        ),
+                                ]),
+                            ]),
                     ]),
-                Builder::make('components')
-                    ->label('Form Elements')
-                    ->addBetweenActionLabel('Insert between elements')
-                    ->columnSpan(2)
-                    ->collapsible()
-                    ->collapsed(true)
-                    ->blockNumbers(false)
-                    ->cloneable()
-                    ->afterStateHydrated(function (Set $set, Get $get) {
-                        Session::put('elementCounter', self::getHighestID($get('components')) + 1);
-                    })
-                    ->blocks([
-                        FormFieldBlock::make(fn() => FormTemplateHelper::calculateElementID()),
-                        FieldGroupBlock::make(fn() => FormTemplateHelper::calculateElementID()),
-                        ContainerBlock::make(fn() => FormTemplateHelper::calculateElementID()),
-                    ]),
-                Actions::make([
-                    Action::make('Generate Form Template')
-                        ->action(function (Get $get, Set $set) {
-                            $formId = $get('id');
-                            $jsonTemplate = \App\Helpers\FormTemplateHelper::generateJsonTemplate($formId);
-                            $set('generated_text', $jsonTemplate);
-                        })
-                        ->hidden(fn($livewire) => ! ($livewire instanceof \Filament\Resources\Pages\ViewRecord)),
-                    Action::make('Preview Form Template')
-                        ->url(function (Get $get) {
-                            $jsonTemplate = $get('generated_text');
-                            $encodedJson = base64_encode($jsonTemplate);
-                            return route('forms.rendered_forms.preview', ['json' => $encodedJson]);
-                        })
-                        ->openUrlInNewTab()
-                        ->disabled(fn(Get $get) => empty($get('generated_text')))
-                        ->hidden(fn($livewire) => ! ($livewire instanceof \Filament\Resources\Pages\ViewRecord)),
-                ]),
-                Textarea::make('generated_text')
-                    ->label('Generated Form Template')
-                    ->columnSpan(2)
-                    ->rows(15)
-                    ->hidden(fn($livewire) => ! ($livewire instanceof \Filament\Resources\Pages\ViewRecord)),
             ]);
     }
